HLSL EvaluateAttributeAtSample emulation for Vulkan (#783)
* HLSL EvaluateAttributeAtSample emulation for Vulkan

* EvaluateAttributeAtSample: Sample value is computed via ddx / ddy on the input parameter multiplied by a hardcoded sample offset.
* Implemented alpha to coverage.
* Removed PipelineCacheKey in ezImageCopyVulkan. The key contained raw pointers which would crash on e.g. shader reload. The values are already cached in ezResourceCacheVulkan so this just replaced one unsafe hashmap lookup with 3 safe hashmap lookups.
* Compile fix in DocumentNodeManager.

* Applied code formatting rules

+++ Code/Tools/Libs/ToolsFoundation/NodeObject/Implementation/DocumentNodeManager.cpp
@@ -531,7 +531,7 @@ bool ezDocumentNodeManager::PasteObjects(const ezArrayPtr<ezDocument::PasteInfo>
       }
     }
 
-    vAvgPos /= nodeCount;
+    vAvgPos /= (float)nodeCount;
     const ezVec2 vMoveNode = -vAvgPos + pickedPosition;
 
     for (const ezDocumentObject* pObject : AddedObjects)
